Added content and subcontent views
