+++ addon/views/llama-subcontent.js
@@ -1,0 +1,9 @@
+import Em from 'ember';
+
+var LlamaSubcontent = Em.CollectionView.extend({
+	classNames: ['llama-subcontent'],
+
+	rows: null
+});
+
+export default LlamaSubcontent;
